test(core): achieve 100% core coverage with professional test distribution
- Redistributed targeted tests from test_coverage_gap_fill.py into respective component test files.
- Deleted the temporary test_coverage_gap_fill.py.
- Achieved literal 100% coverage for all core engine modules.
- Refined formatting and resolved lint/type-checking warnings.
- Verified all Justfile developer recipes pass.
