--- tests/unit/core/test_coverage_gap_fill.py
@@ -1,343 +1,0 @@
-import unittest.mock as mock
-import xml.etree.ElementTree as ET
-
-import pytest
-from linkforge_core.base import RobotParserError
-from linkforge_core.generators.xacro_generator import XACROGenerator
-from linkforge_core.models import (
-    Box,
-    Color,
-    Joint,
-    JointType,
-    Link,
-    Material,
-    Robot,
-    Vector3,
-    Visual,
-)
-from linkforge_core.parsers.urdf_parser import URDFParser
-from linkforge_core.parsers.xacro_parser import XacroResolver
-
-
-class TestCoverageGapFill:
-    """Tests designed specifically to fill coverage gaps."""
-
-    def test_generator_generate_macro_definition_no_joint(self):
-        """Cover line 534 in xacro_generator.py."""
-        gen = XACROGenerator()
-        root = ET.Element("robot")
-        # Template list with link but None joint
-        link = Link(name="l")
-        group = [(link, None)]  # type: ignore
-        # Should return early and not crash
-        gen._generate_macro_definition(root, "sig", group)
-        assert len(root) == 0
-
-    def test_generator_generate_macro_call_no_joint(self):
-        """Cover line 611 in xacro_generator.py."""
-        gen = XACROGenerator()
-        root = ET.Element("robot")
-        link = Link(name="l")
-        # Should return early
-        gen._generate_macro_call(root, "sig", link, None)
-        assert len(root) == 0
-
-    def test_generator_visual_name(self):
-        """Cover line 635 in xacro_generator.py."""
-        gen = XACROGenerator()
-        robot = Robot(name="test")
-        link = Link(name="l")
-        vis = Visual(geometry=Box(size=Vector3(1.0, 1.0, 1.0)), name="my_visual")
-        link.visuals.append(vis)
-        robot.add_link(link)
-
-        xml = gen.generate(robot)
-        assert 'visual name="my_visual"' in xml
-
-    def test_generator_write_no_split(self, tmp_path):
-        """Cover line 759 in xacro_generator.py."""
-        gen = XACROGenerator(split_files=False)
-        robot = Robot(name="test")
-        robot.add_link(Link(name="base"))
-        out = tmp_path / "out.xacro"
-        gen.write(robot, out)
-        assert out.exists()
-
-    def test_generator_split_files_with_macros(self, tmp_path):
-        """Cover lines 790-841 in xacro_generator.py."""
-        gen = XACROGenerator(split_files=True, generate_macros=True)
-        # Create robot with repeated geometry to trigger macros
-        robot = Robot(name="test")
-
-        # Base link
-        base = Link(name="base")
-        robot.add_link(base)
-
-        # Link 1
-        l1 = Link(name="l1")
-        v1 = Visual(geometry=Box(size=Vector3(1.0, 1.0, 1.0)))  # Identical box
-        l1.visuals.append(v1)
-        robot.add_link(l1)
-
-        # Link 2
-        l2 = Link(name="l2")
-        v2 = Visual(geometry=Box(size=Vector3(1.0, 1.0, 1.0)))  # Identical box
-        l2.visuals.append(v2)
-        robot.add_link(l2)
-
-        # Joints needed for macro grouping
-        j1 = Joint(name="j1", type=JointType.FIXED, parent="base", child="l1")
-        robot.add_joint(j1)
-        j2 = Joint(name="j2", type=JointType.FIXED, parent="base", child="l2")
-        robot.add_joint(j2)
-
-        out = tmp_path / "main.xacro"
-        gen.write(robot, out)
-
-        assert out.exists()
-        macros_file = tmp_path / "test_macros.xacro"
-        macros_content = macros_file.read_text()
-        assert 'macro name="box_' in macros_content
-        assert '_macro"' in macros_content
-        assert 'include filename="test_macros.xacro"' in out.read_text()
-
-    def test_parse_gazebo_element_with_plugins(self):
-        """Cover line 991 in urdf_parser.py."""
-        parser = URDFParser()
-        xml = """
-        <robot name="r">
-            <gazebo>
-                <plugin name="p" filename="lib.so"/>
-            </gazebo>
-        </robot>
-        """
-        robot = parser.parse_string(xml)
-        assert len(robot.gazebo_elements) == 1
-        assert len(robot.gazebo_elements[0].plugins) == 1
-        assert robot.gazebo_elements[0].plugins[0].name == "p"
-
-    def test_parse_file_iterparse_error(self, tmp_path):
-        """Cover line 1232 in urdf_parser.py."""
-        path = tmp_path / "test.urdf"
-        path.touch()
-        parser = URDFParser()
-
-        # Mock ET.iterparse to raise ParseError
-        with (
-            mock.patch("xml.etree.ElementTree.iterparse", side_effect=ET.ParseError("Bad XML")),
-            pytest.raises(RobotParserError, match="Failed to parse URDF XML"),
-        ):
-            parser.parse(path)
-
-    def test_joint_renaming_collision_loop(self):
-        """Cover line 1336 in urdf_parser.py."""
-        # Need to simulate: add 'j' -> exists. try 'j_dup_1' -> exists. try 'j_dup_2' -> ok.
-        robot = Robot(name="test")
-        parser = URDFParser()
-        joint = Joint(name="j", type=JointType.FIXED, parent="p", child="c")
-        elem = ET.Element("joint", name="j")
-
-        with mock.patch.object(robot, "add_joint") as m:
-            m.side_effect = [
-                ValueError("already exists"),  # j exists
-                ValueError("already exists"),  # j_duplicate_1 exists
-                None,  # j_duplicate_2 ok
-            ]
-
-            parser._add_joint_robust(robot, joint, elem)
-
-            assert m.call_count == 3
-            # verify call args if needed, but strict order is enough for cover
-
-    def test_resolve_file_flatten_container(self, tmp_path):
-        """Cover line 122 in xacro_parser.py via resolve_file."""
-        resolver = XacroResolver()
-        path = tmp_path / "test.xacro"
-        # Content that resolves to a container via a conditional
-        path.write_text("""
-        <robot xmlns:xacro="http://www.ros.org/wiki/xacro">
-            <xacro:if value="1">
-                <link name="l"/>
-            </xacro:if>
-        </robot>
-        """)
-
-        # resolve_file -> _finalize_urdf -> _append_filtered
-        # The container from if should be flattened.
-        xml = resolver.resolve_file(path)
-        assert '<link name="l"' in xml
-
-    def test_append_filtered_single_element(self):
-        """Cover line 117 in xacro_parser.py."""
-        # Create a dummy class to access the inner function?
-        # No, it's defined inside resolve_file.
-        # Check if we can trigger it.
-        # It's called with `list(container)` or `list(item)`.
-        # `list()` always returns a list.
-        # So passing an Element is impossible in current code structure unless I pass an Element assuming it's a list (which would fail iteration).
-        # Actually, wait.
-        # def _append_filtered(..., items: list[ET.Element] | ET.Element):
-        # The code supports passing a single Element.
-        # But existing calls wrap in list().
-        # So it's dead code for coverage unless I mock-call it?
-        # I cannot easily mock an inner function.
-        # But wait, if I can make `list(item)` return an Element? No.
-        pass
-
-    def test_handle_include_missing_file_warning(self):
-        """Cover line 257 in xacro_parser.py."""
-        resolver = XacroResolver()
-        xml = ET.fromstring(
-            '<xacro:include xmlns:xacro="http://www.ros.org/wiki/xacro" filename="missing.xacro"/>'
-        )
-
-        with mock.patch("linkforge_core.parsers.xacro_parser.logger") as m:
-            resolver.resolve_element(xml)
-            assert m.warning.called
-            assert "Could not find included file" in m.warning.call_args[0][0]
-
-    def test_property_block_assignment(self):
-        """Cover line 236 in xacro_parser.py."""
-        resolver = XacroResolver()
-        xml = ET.fromstring("""
-        <xacro:property xmlns:xacro="http://www.ros.org/wiki/xacro" name="block">
-            <child/>
-        </xacro:property>
-        """)
-        resolver.resolve_element(xml)
-        assert "block" in resolver.properties
-        assert isinstance(resolver.properties["block"], list)
-
-    def test_handle_load_json_module_missing(self, tmp_path):
-        """Cover line 547-548 in xacro_parser.py."""
-        from linkforge_core.parsers.xacro_parser import XacroResolver
-
-        resolver = XacroResolver(start_dir=tmp_path)
-        with mock.patch("linkforge_core.parsers.xacro_parser.json", None):
-            res = resolver._handle_load_json("foo.json")
-            assert res == {}
-
-    def test_find_common_prefix_empty(self):
-        """Cover line 402 in xacro_generator.py."""
-        gen = XACROGenerator()
-        assert gen._find_common_prefix([]) == ""
-
-    def test_generator_extract_material_property_implementation(self):
-        """Cover lines 670-671 in xacro_generator.py - Implementation."""
-        gen = XACROGenerator(extract_materials=True, advanced_mode=True)
-        # robot = Robot(name="test") - Removed unused variable
-
-        # Manually populate material_properties to simulate extraction having happened
-        # (or rely on robust integration test)
-        # Using manual population is easier for unit test of _add_material_element
-        gen.material_properties["blue"] = "color_blue"
-
-        root = ET.Element("link")
-        mat = Material(name="blue", color=Color(0.0, 0.0, 1.0, 1.0))
-
-        gen._add_material_element(root, mat)
-
-        xml = ET.tostring(root).decode()
-        assert 'rgba="${color_blue}"' in xml
-
-    def test_link_renaming_robustness(self):
-        """Cover lines 1302-1303 in urdf_parser.py."""
-        robot = Robot(name="test")
-        parser = URDFParser()
-        link = Link(name="l")
-
-        # Mock robot.add_link to fail twice then succeed
-        with mock.patch.object(robot, "add_link") as m:
-            m.side_effect = [
-                ValueError("Link 'l' already exists"),
-                ValueError("Link 'l_duplicate_1' failed"),  # Trigger exception inside loop
-                None,  # Succeeds for l_duplicate_2
-            ]
-
-            with mock.patch("linkforge_core.parsers.urdf_parser.logger") as mock_logger:
-                parser._add_link_robust(robot, link)
-                assert m.call_count == 3
-                assert mock_logger.warning.called
-                assert "Renamed duplicate link" in mock_logger.warning.call_args[0][0]
-
-    def test_handle_load_yaml_error(self, tmp_path):
-        """Cover line 540-542 in xacro_parser.py."""
-        resolver = XacroResolver(start_dir=tmp_path)
-        path = tmp_path / "bad.yaml"
-        path.touch()
-
-        with (
-            mock.patch(
-                "linkforge_core.parsers.xacro_parser.yaml.safe_load",
-                side_effect=Exception("Read error"),
-            ),
-            mock.patch("linkforge_core.parsers.xacro_parser.logger") as mock_logger,
-        ):
-            res = resolver._handle_load_yaml("bad.yaml")
-            assert res == {}
-            assert mock_logger.error.called
-
-    def test_handle_load_json_error(self, tmp_path):
-        """Cover line 557-559 in xacro_parser.py."""
-        resolver = XacroResolver(start_dir=tmp_path)
-        path = tmp_path / "bad.json"
-        path.touch()
-
-        with (
-            mock.patch(
-                "linkforge_core.parsers.xacro_parser.json.load", side_effect=Exception("Read error")
-            ),
-            mock.patch("linkforge_core.parsers.xacro_parser.logger") as mock_logger,
-        ):
-            res = resolver._handle_load_json("bad.json")
-            assert res == {}
-            assert mock_logger.error.called
-
-    def test_substitute_mixed_text(self):
-        """Cover lines 491-493, 498 in xacro_parser.py."""
-        resolver = XacroResolver()
-        resolver.properties["p"] = 1.5
-        res = resolver._substitute("val=${p}m")
-        assert res == "val=1.5m"
-
-    def test_cleanup_non_string_tag(self):
-        """Cover lines 578-579 in xacro_parser.py."""
-        resolver = XacroResolver()
-        # Create element with a Comment child (tag is a function/type, not string)
-        root = ET.Element("root")
-        comment = ET.Comment("test")
-        root.append(comment)
-
-        # Mock serialize_xml to avoid crash if finalize tries to serialize it
-        with mock.patch("linkforge_core.utils.xml_utils.serialize_xml", return_value=""):
-            resolver._finalize_urdf(root)
-            # Should not crash.
-
-    def test_try_parse_typed_value_yaml_error(self):
-        """Cover lines 455-456 in xacro_parser.py."""
-        resolver = XacroResolver()
-        # Mock yaml.safe_load to raise Exception
-        with mock.patch(
-            "linkforge_core.parsers.xacro_parser.yaml.safe_load", side_effect=Exception("fail")
-        ):
-            res = resolver._try_parse_typed_value("foo")
-            assert res == "foo"
-
-    def test_find_file_package_uri(self, tmp_path):
-        """Cover line 601 in xacro_parser.py."""
-        resolver = XacroResolver(start_dir=tmp_path)
-        with mock.patch("linkforge_core.parsers.xacro_parser.resolve_package_path") as m:
-            m.return_value = tmp_path / "resolved.urdf"
-            res = resolver._find_file("package://my_pkg/test.urdf")
-            assert res == tmp_path / "resolved.urdf"
-            assert m.called
-
-    def test_find_common_prefix_internal_logic(self):
-        """Ensure _find_common_prefix logic is fully covered."""
-        gen = XACROGenerator()
-        # Test basic cases
-        assert gen._find_common_prefix(["arm_link", "arm_joint"]) == "arm"
-        assert gen._find_common_prefix(["fl_wheel", "fr_wheel"]) == "wheel"
-
-        # Trigger line 402 again explicitly
-        assert gen._find_common_prefix([]) == ""
